Fix autoindexes treated as phase 1 content [RHELDST-23816]
The automatically generated directory listings are a form of entry point
and ought to be treated as phase 2 content, like the rest of the entry
points. If this is not done, it means that (e.g.) yum repodata and the
directory listings generated from it are not guaranteed to be committed
together, and so may sometimes become out of sync with each other.

Tweak the publish code to treat these files in the same way as any other
entry point file.