Enforce /origin/files invariant [RHELDST-23443]
In order to prevent conflicts, ensure that all clients follow the
correct usage of the /origin/files/sha256 structure.

When clients publish underneath the /origin/files/sha256/ path, it
is enforced that:

(1) The path always matches this format:
/origin/files/sha256/(first two letters of sha256sum)/(full sha256sum)/(basename)

(2) Every object_key is either absent, or is equal to the full sha256sum
present in the path.

When either rule is broken, the request is denied with a 400 response.