Keep a record of published paths
When we commit a publish, for any published phase2 content, upsert a
record into a new published_paths table. exodus-gw will keep a
lightweight record of the entry points it has written. Like other data,
anything older than a certain threshold is automatically cleaned up.

This commit alone has no user impact. It is intended for later usage
in RHELDST-23300 (cache flushing on alias update). For example:
if deploy-config is used to update the /content/dist/rhel9/9/ alias to
point to a new destination, we can use this table to find all formerly
published paths matching "/content/dist/rhel9/9/*" as we know that these
now need to have cache flushed.