Flush affected subtrees during alias updates [RHELDST-23300]
When an alias is updated, we should automatically flush cache for any
known phase2 content underneath that alias. This commit implements that
behavior, using the recently added PublishedPath model.

For example, if /content/dist/rhel8/8 was pointing at a certain rhel8
version and was then updated to point to another, the content of every
yum repo underneath that path has potentially changed. We should
therefore flush cache for any repos underneath that path.

(Previously, rhsm-update-releasever attempted to do this, but only
exodus-gw can do it accurately since only exodus-gw knows exactly which
paths have been updated on the CDN.)