cache: skip non-kickstart treeinfo [RHELDST-24308]
It turns out there is a special case in the current CDN config forcing
a fast 404 response on treeinfo URLs not ending in "/kickstart/treeinfo".
A side-effect of forcing that response is that any requests to cache
flush on such a path will fail, e.g.

    flush /some/kickstart/treeinfo => OK

    flush /some/treeinfo => always fails

As such, we'll need a special case to filter out these paths from any
attempted cache flushes. There is never any need for cache flush if the
response is forced to 404 anyway.

Note this is unlikely to come up in any real publishes of kickstart
repos since those are always using /kickstart in the path.
It came up in the context of migrating some old content to exodus-cdn,
because some files named "treeinfo" exist underneath /origin/files which
will trigger the usual cache flush attempts.