Make it possible to bypass /origin/files policy [RHELDST-22253]
RHELDST-23443 started to enforce that content published under
/origin/files must abide by the established naming scheme (e.g. checksum
in path must match checksum of content).

Problem: content from legacy storage is still being migrated via
exodus-gw, and some of that content fails to comply. We should permit
migration of such content as-is.

As there is a use-case for bypassing these checks, but only for the user
performing the migration, add a new role supporting this. If the calling
user has e.g. "live-ignore-policy", they will be permitted to bypass
this specific check in "live". This will be granted to the user
performing the migration.

The relevant code was rewritten a bit and placed next to the other
validation code so that it's cleaner to catch/ignore the exception when
needed.