mypy fixes for get_current_message()
With respect to the CurrentMessage.get_current_message() function from
dramatiq middleware, either mypy or relevant type hints have become
more accurate and now reflect that the function is allowed to return
None when no message is being handled (the return type now being
"Message[Any] | None").

It is a precondition of this code that a message is being handled (via
actor invocation). Add asserts to this effect, which will let mypy rule
out the None case.