Write compressed configuration [RHELDST-25461]
Start writing the configuration in compressed binary form.

For more info on the motivation, see the corresponding change in
exodus-lambda:
https://github.com/release-engineering/exodus-lambda/pull/586

Regarding the test updates in this commit, the existing fixture used to
mock the DynamoDB query for config was parametrized so that existing
tests would cover both binary and non-binary config cases. However,
since that was previously a global autouse fixture, that would have
caused every test to run twice even if the test did not deal with the
config at all.

To avoid this, fixtures were split into a global autouse boto session
fixture (ensuring all tests still avoid any real AWS calls) and a
non-autouse fixture for the DynamoDB client. The new fixture had to be
explicitly added to certain tests.