revert (partial): write compressed configuration [RHELDST-25461]
This commit partially reverts db4a3443feadd74cb; specifically, it
disables the writing of compressed config and reverts back to writing
JSON config as before.

This is being reverted due to an issue noticed during testing. It seems
that the written items are ending up with *two* layers of base64
encoding, which is not intended.
The boto docs[1] use base64 strings as example arguments, giving the
impression the caller is expected to take care of base64 encoding, but
in fact botocore internally does the encoding; if the client also
encodes, we end up with two layers of encoding.

There is also a bug filed relating to this[2].

The code here still seems to "work" since the same mistake is made
on both the writing and reading end, but the goal is to make the
config smaller and having double-encoding works against that.

It should be easy enough to fix, but I'd like some time to confirm my
understanding of how it works, check whether exodus-lambda needs a fix
and also check whether localstack and AWS are behaving the same.
Hence I'll revert this for now and keep writing the old style of config.

[1] https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/dynamodb/client/put_item.html
[2] https://github.com/aws/aws-cli/issues/1097