Exclude empty exclude_paths from config [RHELDST-29250]
Currently, when we attempt to use exodus-gw's "get config" endpoint
to fetch the current config stored in DynamoDB, exodus-gw sanitizes
the "get config" endpoint's response by including an "exclude_paths"
property on all aliases, even when the property is empty.

```
{
  [...]
  "origin_alias": [
    {
      "src": "/content/origin",
      "dest": "/origin",
      "exclude_paths": []
    },
    {
      "src": "/origin/rpm",
      "dest": "/origin/rpms",
      "exclude_paths": []
    }
  ],
  [...]
  "rhui_alias": [
  [...]
    {
      "src": "/content/beta/rhel/rhui",
      "dest": "/content/beta/rhel",
      "exclude_paths": []
    },
  [...]
}
```

The config POSTed to exodus-gw is generated by exodus-config; the
generated config does not enforce adding an exclude_paths property on
all aliases; exodus-config only appends an exclude_paths property onto
an alias when the exclude_paths property is populated/non-empty.

So, when generating a diff between the new config (generated by
exodus-config) and the current config fetched from exodus-gw, the diff
will always suggest that the new config is attempting to remove the
empty "exclude_paths" property from all aliases:

```
--- config from exodus-gw GET config endpoint
+++ config created by exodus-config
@@ -3861,277 +3861,362 @@
     var: releasever
 origin_alias:
 - dest: /origin
-  exclude_paths: []
   src: /content/origin
 - dest: /origin/rpms
-  exclude_paths: []
   src: /origin/rpm
[...]
```

This diff is misleading; one may think that the config stored in
DynamoDB contains an exclude_paths property on each alias, but the
config stored in DynamoDB is identical to the config produced by
exodus-config (meaning that the aliases stored in the config in
DynamoDB do not contain any empty exclude_paths properties):

```
{
  "origin_alias": [
    {
      "src": "/content/origin",
      "dest": "/origin",
    },
    {
      "src": "/origin/rpm",
      "dest": "/origin/rpms",
    }
  ],
  [...]
  "releasever_alias": [
  [...]
  {
    'src': '/content/dist/rhel8/8',
    'dest': '/content/dist/rhel8/8.10',
    'exclude_paths': [
      '/files/',
      '/iso/'
    ],
  },
  [...]
}
```

The "exclude_paths" property was designed to be an optional property
added on to aliases. It is not expected by exodus-lambda, so there is
no reason as to why the "get config" endpoint's responses should be
sanitized such that an empty "exclude_paths" property is always
appended to all aliases without a populated "exclude_paths" property.

This commit adjusts the response from the "get config" endpoint such
that empty "exclude_paths" properties are excluded from responses.