Revert "Mirror release version alias writes [RHELDST-28333]"
This reverts commit 3fd938078d40a41c88fe3adc1fc067432e7c28ba.

When publishing items for paths with a releasever alias in effect,
mirroring causes us to write to twice the number of items (because
exodus-gw writes to the original releasever path, as well as the path
with the resolved releasever alias).

AWS imposes a hard limit of 25 items per batch write. exodus-gw
partitions the items into appropriate batch sizes within the publish.
But, because the extra mirrored items are added to the batch outside
of the publish (immediately before writing the items to DynamoDB),
exodus-gw attempts to include these extra items within a pre-partitioned
batch (i.e., a batch which has already been resized to contains <=25
items). So adding additional items into a pre-partitioned batch will
exceed the limit of 25 items per batch write, which causes the batch
write to fail.