multipart upload: make abort response same as complete [RHELDST-30644]
Previously, when a duplicate upload was detected when trying to complete
a multipart upload, an empty reponse was returned as in
`abort_multipart_upload`. However, s3 apis expect an xml response. For
compatibility and ease, this commit changes the response to the same as
that produced by `complete_multipart_upload`, minus a few keys that
are only available by the real CompleteMultipartUpload. This keeps
things as "same" as they can possibily be while avoiding uploading
duplicates/overwriting.