Config deployment should flush resolved paths [RHELDST-32942]
Problem: exodus-gw was expected to have triggered an automatic cache
flush of the mainline kickstart paths after provisioning the CDN (i.e.,
after running a config deployment containing updated releasever
aliases). This did not happen. This is because post-config deployment,
exodus-gw only flushed paths that were prefixed by the alias src (i.e.,
the mainline repos) found in the PublishedPath db (which consists of
recently published paths). The kickstart repos are only published to
the dest side of the alias (i.e., DOT repos), so the kickstart paths will
never appear in the PublishedPath db when we query for src paths.

To account for this, the post-deployment automatic cache flush behavior
must flush paths on the "src" side of the alias, even if the PublishedPath
db only contains paths on the "dest" side of the alias (as the content will
need to be made available on the "src" side of the alias). This means
that the post-deployment automatic cache flush behavior must resolve any
paths on the "dest" side of the alias into the corresponding "src" path,
and flush that path.

For example, if a config deployment updates the `/content/dist/rhel9/9` alias
to point to the next minor RHEL version, and we recently published to the
following paths:

```
/content/dist/rhel9/9.7/x86_64/kickstart/treeinfo
/content/dist/rhel9/9.7/x86_64/os/Packages/a/
/content/dist/rhel9/9.7/x86_64/repodata/repomd.xml
```

The following paths should be flushed during the config deployment
(because the content published at these paths will have changed after
the config deployment):

```
/content/dist/rhel9/9/x86_64/kickstart/treeinfo
/content/dist/rhel9/9/x86_64/os/Packages/a/
/content/dist/rhel9/9/x86_64/repodata/repomd.xml
```

The following paths should not be flushed (because the content published
at these paths will have not changed after the config deployment):

```
/content/dist/rhel9/9.7/x86_64/kickstart/treeinfo
/content/dist/rhel9/9.7/x86_64/os/Packages/a/
/content/dist/rhel9/9.7/x86_64/repodata/repomd.xml
```