cache: Skip hidden treeinfo [RHELDST-24308, RHELDST-34594]
It turns out there is a special case in the current CDN config forcing
a fast 404 response on URLs ending in "/.treeinfo". A side-effect of
forcing that response is that any requests to cache flush on such a
path will fail, e.g.

    flush /some/kickstart/treeinfo => OK

    flush /some/kickstart/.treeinfo => always fails

As such, we'll need a special case to filter out this path from any
attempted cache flushes. There is never any need for cache flush if the
response is forced to 404 anyway.

Relates to: 6e53fc91fb7a61ac6484c